pyproject.toml
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "thanos"
version = "0.0.0"

[tool.poetry.dependencies]
anyio = "^4.2.0"
asyncache = "^0.3.1"
authlib = "^1.3.0"
cachetools = "^5.3.2"
fastapi = "^0.109.0"
httpx = {extras = ["brotli", "http2"], version = "^0.26.0"}
itsdangerous = "^2.1.2"
jinja2 = "^3.1.3"
motor = "^3.3.2"
orjson = "^3.9.10"
psutil = "^5.9.7"
python = "^3.12"
python-multipart = "^0.0.6"
pyyaml = "^6.0.1"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
xmltodict = "^0.13.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
]

indent-width = 4
line-length = 120
target-version = "py312"

[tool.ruff.lint]
ignore = [
  "ARG001", # unused-function-argument
  "B008", # function-call-in-default-argument
  "S101", # assert
  "S108", # hardcoded-temp-file
]
select = [
  # see https://docs.astral.sh/ruff/rules/ for rules documentation
  "A", # flake8-builtins
  "ARG", # flake8-unused-arguments
  "ASYNC", # flake8-async
  "B", # flake8-bugbear
  # "COM", # flake8-commas
  "C4", # flake8-comprehensions
  "E4", # pycodestyle
  "E7",
  "E9",
  "F", # pyflakes
  # "FBT", # flake8-boolean-trap
  "FLY", # flynt
  # "FURB", # refurb (preview)
  "G", # flake8-logging-format
  "I", # isort
  "INT", # flake8-gettext
  # "LOG", # flake8-logging (preview)
  "N", # pep8-naming
  "NPY", # numpy
  "UP", # pyupgrade
  "Q", # flake8-quotes
  "PERF", # perflint
  "PGH", # pygrep-hooks
  "PIE", # flake8-pie
  # "PL", # pylint
  "PT", # flake8-pytest-style
  "PTH", # flake8-use-pathlib
  "PYI", # flake8-pyi
  "RSE", # flake8-raise
  "RUF", # ruff
  "S", # flake8-bandit
  "SIM", # flake8-simplify
  "SLF", # flake8-self
  "SLOT", # flake8-slots
  "T10", # flake8-debugger
  # "T20", # flake8-print
  # "TRY", # tryceratops
  "YTT", # flake8-2020
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "single"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "lf"

[tool.ruff.lint.flake8-builtins]
builtins-ignorelist = ["id", "open", "type", "filter", "format"]

[tool.ruff.lint.flake8-gettext]
extend-function-names = ["pgettext", "npgettext"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"
multiline-quotes = "double"

[tool.ruff.lint.pep8-naming]
extend-ignore-names = ["ET"]

[tool.ruff.lint.pylint]
max-args = 10
